Decompose the 1403-line AnnotationWorkspace component into three cohesive hooks under components/annotation/hooks (useAnnotationDialogs for the dialog and transcript state, useSummaryFlow for the claim timestamp-capture and draft-claim flow, and useAnnotationState for the playback state, keyframe and edit handlers, and autosave setup), leaving the component a thin composition of the hooks plus its render with state, effects, dependency arrays, handler logic, and output unchanged.
