+++ CHANGELOG.md
@@ -86,7 +86,9 @@ The 0.5.0 cycle delivers the architecture-modularization roadmap (`notes/archite
 
 - Demo mode (`FOVEA_DEMO_MODE`) exposes system-seeded content to anonymous/all users on the public-demo/booth deployment by widening read access; that widening was duplicated ad hoc across `routes/annotations.ts`, `routes/summaries.ts`, `routes/persona-preferences.ts`, `services/persona-service.ts`, `services/video-access-service.ts`, and `services/world-state-service.ts`, each with a different shape, so what demo mode exposed was hard to audit. It is now consolidated into one module, `server/src/lib/demo-rbac.ts` — the single source of truth for the per-subject demo read scopes (`{ source: { startsWith: 'demo-fixture' } }` for annotations, `{ isSystemGenerated: true }` for personas) and the demo read predicates (videos, summaries, world state, system-persona access). Each helper gates on `isDemoModeEnabled()` internally, so no route or service scatters that flag check anymore (a grep of `routes/` + `services/` for `isDemoModeEnabled` is now empty). Pure consolidation: every site's effective query/decision is byte-for-byte preserved — demo mode exposes the same content, and with demo off a self-hoster's per-user RBAC is exactly as before.
 
-#### Reviewers Can Resolve Cross-User Object Names
+#### Annotation Workspace Logic Extracted into Hooks
+
+- The 1403-line `annotation-tool/src/components/annotation/AnnotationWorkspace.tsx` is decomposed into three cohesive custom hooks under `components/annotation/hooks/`: `useAnnotationDialogs` (the editor/summary/detection/transcript dialog state and the transcript-request flow), `useSummaryFlow` (the claim timestamp-capture wiring and the draft-claim auto-open), and `useAnnotationState` (the playback state, the keyframe/edit handlers, and the auto-save setup). The component drops to a thin composition of the three hooks plus its render. Pure move — state, effects, dependency arrays, handler logic, and render output are unchanged (the keyframe handlers still pass the freshly-mutated array to `forceSave`, and the auto-save change-detection snapshot is byte-identical), so behavior is preserved.
 
 - The annotations list endpoint (`GET /api/annotations/:videoId`) now returns an optional server-resolved `linkedObjectName`, and `AnnotationOverlay` uses it as a fallback badge label. World objects (entities/events/times/locations) are per-user, so a reviewer reading another annotator's object annotation previously saw a generic "Entity" badge — their own world could not resolve the other annotator's object id. The server now resolves the linked object's display name from the **owner's** world (batched: one `worldState.findMany` over the distinct owners on the page), gated implicitly by the existing `accessibleBy(read)` filter so only annotations the caller may already read are resolved and only the name is exposed. The frontend still prefers the live local object when the caller shares the world; `linkedObjectName` is the fallback when it cannot.
